fix: currency input limit
